+++ test/correctness/PySys_internal_185/Input/Test3/pysystest.py
@@ -1,0 +1,20 @@
+__pysys_title__   = r""" My test """ 
+#                        ================================================================================
+__pysys_purpose__ = r""" """ 
+	
+__pysys_authors__ = "bsp"
+__pysys_created__ = "2022-03-11"
+#__pysys_skipped_reason__   = "Skipped until Bug-1234 is fixed"
+
+import pysys
+from pysys.constants import *
+
+import os, sys, math, shutil, glob
+
+class PySysTest(pysys.basetest.BaseTest):
+
+	def execute(self):
+		pass
+		
+	def validate(self):
+		pass